Stop file watcher before starting to process events
This avoids the ConcurrentModificationException we've seen sometimes pop up.
