+++ subprojects/core/src/main/java/org/gradle/internal/vfs/DarwinFileWatcherRegistry.java
@@ -17,7 +17,6 @@
 package org.gradle.internal.vfs;
 
 import net.rubygrapefruit.platform.Native;
-import net.rubygrapefruit.platform.file.FileWatcher;
 import net.rubygrapefruit.platform.internal.jni.OsxFileEventFunctions;
 import org.gradle.internal.vfs.impl.WatchRootUtil;
 import org.gradle.internal.vfs.impl.WatcherEvent;
@@ -26,22 +25,16 @@ import org.gradle.internal.vfs.watch.FileWatcherRegistryFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
-public class DarwinFileWatcherRegistry implements FileWatcherRegistry {
+public class DarwinFileWatcherRegistry extends AbstractEventDrivenFileWatcherRegistry {
     private static final Logger LOGGER = LoggerFactory.getLogger(DarwinFileWatcherRegistry.class);
 
-    private final FileWatcher watcher;
-    private final List<WatcherEvent> events = new ArrayList<>();
-
     public DarwinFileWatcherRegistry(Set<Path> watchRoots) {
-        this.watcher = Native.get(OsxFileEventFunctions.class)
+        super(events -> Native.get(OsxFileEventFunctions.class)
             .startWatching(
                 watchRoots.stream()
                     .map(Path::toString)
@@ -49,17 +42,7 @@ public class DarwinFileWatcherRegistry implements FileWatcherRegistry {
                 // TODO Figure out a good value for this
                 300, TimeUnit.MICROSECONDS,
                 (type, path) -> events.add(WatcherEvent.createEvent(type, path))
-            );
-    }
-
-    @Override
-    public void stopWatching(ChangeHandler handler) throws IOException {
-        WatcherEvent.dispatch(events, handler);
-    }
-
-    @Override
-    public void close() throws IOException {
-        watcher.close();
+            ));
     }
 
     public static class Factory implements FileWatcherRegistryFactory {
